+++ myBackupReport.ps1
@@ -1,20 +1,25 @@
-# requires -version 5.1
+#requires -version 5.1
+#requires -module PSScriptTools
 
+#myBackupReport.ps1
+
+# this script uses Format-Value from the PSScriptTools module.
 [cmdletbinding()]
 
 Param(
     [Parameter(Position = 0, HelpMessage = "Enter the path where the backup files are stored.")]
     [ValidateNotNullOrEmpty()]
-    [ValidateScript({Test-Path $_ })]
+    [ValidateScript( { Test-Path $_ })]
     #This is my NAS device
     [string]$Path = "\\ds416\backup",
+    
     [Parameter(HelpMessage = "Get backup files only with no formatted output.")]
     [Switch]$Raw
 )
 
-$reportVer = "1.1"
+$reportVer = "1.2.0"
 
-Write-Verbose "Starting $($myinvocation.mycommand) ver. $ReportVer"
+Write-Verbose "Starting $($myinvocation.mycommand) v.$ReportVer"
 <#
  A regular expression pattern to match on backup file name with named captures
  to be used in adding some custom properties. My backup names are like:
@@ -30,14 +35,14 @@ Write-Verbose "Starting $($myinvocation.mycommand) ver. $ReportVer"
 I am doing so 'pre-filtering' on the file extension and then using the regular
 expression filter to fine tune the results
 #>
-$files = Get-ChildItem $Path\*.zip,$Path\*.rar | Where-Object {$rx.IsMatch($_.name) }
+$files = Get-ChildItem $Path\*.zip, $Path\*.rar | Where-Object { $rx.IsMatch($_.name) }
 Write-Verbose "Found $($files.count) files in $Path"
 
-#add some custom properties to be used with formatted results based on named captures
 foreach ($item in $files) {
     $setpath = $rx.matches($item.name).groups[4].value
     $settype = $rx.matches($item.name).groups[5].value
 
+    #add some custom properties to be used with formatted results based on named captures
     $item | Add-Member -MemberType NoteProperty -Name SetPath -Value $setpath
     $item | Add-Member -MemberType NoteProperty -Name SetType -Value $setType
 }
@@ -47,24 +52,30 @@ if ($raw) {
 }
 else {
     Write-Verbose "Preparing report data"
-    $files | Sort-Object SetPath, SetType, LastWriteTime | Format-Table -GroupBy SetPath -Property LastWriteTime, Length, Name
-    $grouped = $files | Group-Object SetPath
-    $summary = foreach ($item in $grouped) {
-        [pscustomobject]@{
-            BackupSet = $item.name
-            Files  = $item.Count
-            Size   = ($item.group | Measure-Object -Property size -sum).sum
-        }
+    Write-Host "$([char]0x1b)[1;4;38;5;216m`nMy Backup Report$([char]0x1b)[0m"
+    $files | Sort-Object SetPath, SetType, LastWriteTime |
+    Format-Table -GroupBy SetPath -Property @{Name = "Created"; Expression = { $_.LastWriteTime } },
+    @{Name = "SizeMB"; Expression = { Format-Value -input $_.Length -unit MB -decimal 0 } },
+    Name
+$grouped = $files | Group-Object SetPath
+$summary = foreach ($item in $grouped) {
+    [pscustomobject]@{
+        BackupSet = $item.name
+        Files     = $item.Count
+        SizeMB    = ($item.group | Measure-Object -Property size -sum -outvariable m).sum | Format-Value -unit MB -Decimal 2
     }
-
-    $total = [PSCustomObject]@{
-        TotalFiles = ($grouped | Measure-Object -property count -sum).sum
-        TotalSizeMB = [math]::round(($summary.size | Measure-Object -sum).sum/1MB,4)
-    }
-    Write-Host "Backup Summary $((Get-Date).ToShortDateString())" -ForegroundColor yellow
-    ($summary | sort-object Size -Descending| Format-Table | Out-String).TrimEnd() | Write-Host -ForegroundColor yellow
-
-    ($total | Format-Table  | Out-String).TrimEnd() | Write-Host -ForegroundColor yellow
 }
 
-"$([char]0x1b)[38;5;216mReport version $reportver$([char]0x1b)[0m"
+$total = [PSCustomObject]@{
+    TotalFiles  = ($summary.files | Measure-Object -sum).sum
+    #($grouped | Measure-Object -property count -sum).sum
+    TotalSizeMB = ($summary.sizeMB | Measure-Object -sum).sum
+    #[math]::round(($summary.size | Measure-Object -sum).sum/1MB,4)
+}
+Write-Host "Backup Summary $((Get-Date).ToShortDateString())" -ForegroundColor yellow
+($summary | Sort-Object Size -Descending | Format-Table | Out-String).TrimEnd() | Write-Host -ForegroundColor yellow
+
+($total | Format-Table | Out-String).TrimEnd() | Write-Host -ForegroundColor yellow
+}
+
+Write-Host "$([char]0x1b)[38;5;216m`nReport version $reportver$([char]0x1b)[0m"